* changed int to int32 in playlist struct declaration * added missing members to windows soundcardPMO.h * #defined min definition in jdw.h * #ifdefed windows for io.h in xinglmc.cpp

+++ base/win32/src/soundcardpmo.cpp
@@ -262,7 +262,7 @@ WriteThis(void *pBuffer,int32 length)
 
 	wavhdr = NextHeader();
 
-	if(length > m_data_size)
+	if(length > (int32)m_data_size)
 		length = m_data_size;
 
 	wavhdr->dwBufferLength = length;
